+++ lib/phDeliver.py
@@ -75,56 +75,3 @@ def SenditOut(verbose,list_KeysList, Processed_Results_List, OutFormat, OutFile,
         return True
     Loggy(verbose,900, "------EXIT: SENDITOUT-----")
 
-
-def FullSenditOut(verbose,Processed_Results_List, OutPath, HTMLHeader, HTMLFooter, HTMLCSS):
-    Loggy(verbose,900, "------ENTER: FULLSENDITOUT-----")
-
-    list_KeysList = ["Title", "Count", "Further Details"]
-    OutFile = "Report.html"
-    Title = "Full Report Details"
-
-    Loggy(verbose,100, "Beginning Output HTML:" + OutFile)
-
-    for entry in Processed_Results_List:
-        filename = entry[2]
-        entry[2] = "<a href=\"" + filename + "\">Details</a>"
-
-    output = str(tabulate(Processed_Results_List, list_KeysList, tablefmt="html"))
-    output = output.replace("&lt;","<")
-    output = output.replace("&gt;",">")
-    output = output.replace("&quot;",'"')
-
-    HTMLCSS_str = ""
-    HTMLHeader_str = ""
-    HTMLFooter_str = ""
-    HTMLPre_str = "<HTML><head>"
-    HTMLMId_str = "</head><Body>"
-    HTMLEnd_str = "</body></html>"
-    if HTMLHeader:
-        with open(HTMLHeader, 'r') as header:
-            HTMLHeader_str = header.read()
-        HTMLHeader_str = ReplaceHTMLReportVars(HTMLHeader_str, Title)
-
-    if HTMLFooter:
-        with open(HTMLFooter, 'r') as footer:
-            HTMLFooter_str = footer.read()
-        HTMLFooter_str = ReplaceHTMLReportVars(HTMLFooter_str, Title)
-
-    if HTMLCSS:
-        with open(HTMLCSS, 'r') as css:
-            HTMLCSS_str = "<style>\n" + css.read() + "\n</style>"
-
-    Loggy(verbose,500, "File Writing " + OutPath + OutFile)
-    output = HTMLPre_str + HTMLCSS_str + HTMLMId_str + HTMLHeader_str + output + HTMLFooter_str + HTMLEnd_str
-    fsys = open(OutPath + OutFile, "w")
-    fsys.write(output)
-    fsys.close
-    Loggy(verbose,100, "Full report written to Report.html")
-    return True
-    Loggy(verbose,900, "------EXIT: FULLSENDITOUT-----")
-
-
-def ReplaceHTMLReportVars(InputStr, Title):
-    sOutPut = InputStr.replace("--------PH_TITLE-------", str(Title))
-    sOutPut = sOutPut.replace("--------PH_DATE-------", str(date.today()))
-    return sOutPut
